Add å/Å and é/É to the text fonts.
This makes the fonts compatible with the most common usage of the Swedish language and, I think, the Finnish language as well. (Å/å is also needed for Danish and Norwegian, but to complete them, Æ/æ and Ø/ø are also needed.)

I had to make the glyphs a bit compressed, and therefore a bit awkward, especially in the 5px font, so feel free to redraw them. They look (a bit) better on the watch, though...