Refactor the actions framework and add scrolling.
The actions framework is now built upon multiple interfaces instead of one abstract class. I thought that was cleaner. :)

Scrolling is made so that the bottom of the current selection and one line after is shown. When the last item is selected a dashed line is drawn below it (so that the button icon, especially in the case of resettable actions, never overwrites the text).