Remove support for garthy's Media Control app.
Garthy built music controls into his watch firmware (based on 0.7.X). With 0.10.X firmware out, my guess is that nobody is using that firmware anymore, and we have no idea if the support is working at all (the button to open the app may well be overridden by MWM now).