Refactor app framework and add a Music controls action.
This commit lays the groundwork for a more usable Application framework (but still needs work, especially for analog watches). Internal Apps can now be run "stand-alone" (I.E. without being an idle page). The normal "next page" button is an "exit" button for stand-alone apps and will return the watch to idle mode.

Using this, I added an action to the Actions app that opens music controls on top of it.

Also cleaned up a few bits of code here and there.