Make it possible to toggle apps as idle pages.
The Actions app now lists all apps that AppManager knows about (excluding itself and unsupported apps) that are not running as idle pages, making it possible to open these stand-alone.

Using a long press on the exit app/next idle page button on app pages (excluding the Actions app), the user can toggle apps between being an idle page and running stand-alone (which mode is indicated with either an X (stand-alone) or the classic next page (idle page) icon for the top right button).

Right now, although toggling the mode for the media player app is equivalent to changing the setting for enabling or disabling the media player app, the setting isn't changed (technically, the setting can be removed).