New setting: Use app buffer for clock-less widget pages.
A new setting for the digital watch that makes MWM use the application watch buffer for drawing widget pages that doesn't show the clock.

This has the useful benefit of always showing the clock and the watch's status icon when the connection is lost (the watch firmware returns to idle mode on lost connection). The drawback is that the idle mode buttons (settings, watch status etc.) doesn't work for the affected pages.

I made the setting default to true, since the clock is quite a central part of a watch. ;)