Remake "Replay Notification" button to a configurable "Quick Button".
For Digital only, since we don't know what button to use for this on the
Analog.

The Quick Button can be configured to either do nothing, replay last
notification or open the Actions app.

Using this, the Actions app is toggleable, but only if it's set as the
Quick Button action.