Make toggling (supported) apps change their settings.
Lessen confusion for users until there's a better app selection framework in place.

When toggling the Music Controls and Actions apps, their settings in preferences will be changed, so that MWM will start with the same set of pages the next time (albeit not guaranteed to be in the same order).

Also a few fixes to keep stand-alone apps updated on screen.